image-processing-package/requirements.txt
matplotlib
pillow
pandas
numpy